# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "App to save and interactive with contacts"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://contacts-upp.onrender.com"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register user",
        "operationId": "registerUser",
        "description": "Register user with name, email and password",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "description": "User's name",
                    "type": "string",
                    "example": "John Smith"
                  },
                  "password": {
                    "description": "User's password",
                    "type": "string",
                    "example": "123123123"
                  },
                  "email": {
                    "description": "Contact's email",
                    "type": "string",
                    "example": "user@gmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a new user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a new user!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/auth"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login user",
        "operationId": "loginUser",
        "description": "Login user with email and password",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "password": {
                    "description": "User's password",
                    "type": "string",
                    "example": "123123123"
                  },
                  "email": {
                    "description": "Contact's email",
                    "type": "string",
                    "example": "user@gmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged user!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/auth"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "auth": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "code"
        ],
        "properties": {
          "name": {
            "description": "User's name",
            "type": "string",
            "example": "John Smith"
          },
          "phoneNumber": {
            "description": "User's password",
            "type": "string",
            "example": "123123123"
          },
          "email": {
            "description": "User's email",
            "type": "string",
            "example": "user@gmail.com"
          },
          "code": {
            "description": "User's code from google",
            "type": "string",
            "example": "4/0AVG7fiRk0nI0Eh9htwaNCfW9xkOaXc4n_ylvuLDtC2_hBFQNpqOf6G2y56Sh_VnKa4-ZsA"
          }
        }
      }
    },
...
  }
}